adds reports component
